# Optimize Reduce function by reusing CVal