# Description
Summarize the change.
Reuse the CVal enclosures instead of creating new ones.
Query = `* | stats sum(http_status) by hobby, http_method`

Performance improvement is a bit tricky in this case. Here the major time taken is during the operation, in the above example it is taken by the addition operation in the Reduce function and this is quite random.

Ran multiple tests for 10 iterations on around 100M events.
The minimum execution time difference observed between before and after change for Reduce function is around 2.5s (75.68s to 72.97s). And same is the case for for the place where we have optimized by resuing the CVal, the difference is about 2s.
In very rare executions, I have observed the original code to perform better than this one. But, it could majorly be due to operation being executed faster (which is a bit random).

For memory
Have seen reduction in memory usage from 7.49GB to 7.44GB.

Fixes #<issue-number> (link all the GitHub issues this addresses)

# Testing
Describe how you tested this code. How can the reviewers reproduce your tests?

# Checklist:
Before marking your pull request as ready for review, complete the following.

- [x] I have self-reviewed this PR.
- [x] I have removed all print-debugging and commented-out code that should not be merged.
- [x] I have added sufficient comments in my code, particularly in hard-to-understand areas.
- [x] I have formatted the code, if applicable. For Go, I have run `goimports -w .`.